Movement Choices and Node Creation
Renamed some parameters to better represent what they mean
Moved 'makeNodes' from main and into write.go and broke it into 3 parts:
InitializeNodeParameters - initializes all random parameters of the node
SetupCSVNodes - sets up nodes if they are going to be handled by the CSV file
SetupRandomNodes - sets up nodes if they are going to be randomly created

added node.Valid to determine if the node is inside the arena. sensor updating is not handled if the node is not in the arena.