Better out of range testing for clusters
Nodes are only reported in the cluster debug file as out of range from their cluster heads if they have taken a reading since they moved out of range (since that is when they would detect that they are out of range). A small update to the clustering energy model accounts for how cluster heads know when a node has left their cluster.